refactor#125 visible button
